+++ server/models/index.js
@@ -1,3 +1,3 @@
-const User = require('./User');
+const User = require('./User.js');
 
 module.exports = { User };
