+++ server/models/temp
@@ -1,0 +1,53 @@
+const { Schema, model } = require('mongoose');
+const bcrypt = require('bcrypt');
+
+// import schema from Movie.js
+
+const userSchema = new Schema(
+  {
+    username: {
+      type: String,
+      required: true,
+      unique: true,
+    },
+    email: {
+      type: String,
+      required: true,
+      unique: true,
+      match: [/.+@.+\..+/, 'Must use a valid email address'],
+    },
+    password: {
+      type: String,
+      required: true,
+    },
+    savedMovies: [String], //Change to movieSchema once ready
+  },
+  {
+    toJSON: {
+      virtuals: true,
+    },
+  }
+);
+
+// hash user password
+userSchema.pre('save', async function (next) {
+  if (this.isNew || this.isModified('password')) {
+    const saltRounds = 10;
+    this.password = await bcrypt.hash(this.password, saltRounds);
+  }
+
+  next();
+});
+
+// validate password
+userSchema.methods.isCorrectPassword = async function (password) {
+  return bcrypt.compare(password, this.password);
+};
+
+userSchema.virtual('savedMovieCount').get(function () {
+  return this.savedMovies.length;
+});
+
+const User = model('User', userSchema);
+
+module.exports = User;
